CMS-14035 Add a failing unit test
If a document has been saved as draft, the only variant is the draft
variant.
If that document is deleted, it's archived and moved to the attic.
Jeroen Hoffman pointed out that a handle with a draft variant in the
attic does not contain any useful information. Carolien de Kort
mentioned that deleting a document that is saved as draft is funcionally
equal to the cancel action. The cancel action deletes the draft and the
handle and does not move it to the attic.

Add a (failing) unit test that asserts the document should be deleted in
case there is only a draft variant.

(cherry picked from commit b5659f9ffc620effab621e55d583b2a34e90ceb6)