CMS-14285 (back port to 14.5) Filter out the prototype locale
Filter out the default template locale("document-type-locale") instead
of lowering the log level or change all prototypes.

Lowering the log level IMO is not an option, because specifying a wrong
locale should result in an error.  The impact of changing all prototypes
is IMO too high for the severity of the issue.

(cherry picked from commit 0ceae1aefa10dbebd03bd037718af27539938c03)