CMS-14381 [Backport 14.5] Support nested choice fields
When we do a lookup of a nested choice field, we don't know which
field was chosen by looking at the provided fieldPath, as all nodes
of a choice-field have the same name (but not the same type).

We do know that the field has to be a compound field, otherwise a
nested field can not exist. For each choice that is a compound, we
try to lookup the remaining path returning the first hit or null.

(cherry picked from commit 26dc5a30ea4ad58c0680e819b82b674df9f5366c)