CMS-14459 [Backport 14.5] Support 'hidden' component meta data in PMA
if the component is hidden. When the field is missing, it means the
component will not be hidden. When the field is there with value true,
the SPA must make sure to not render the component (in case of a brSM /
ecommerce component make sure not to render the brSM / ecommerce
component)

Eg:

"meta" : {
  "hidden" : true
  "paramsInfo" : {
    "header" : "Yes My Header"
  },
  "params" : { }
}

(cherry picked from commit 07b7e151ae3296c3dbfa803cd6dce8e5b0324523)
(cherry picked from commit 798cb7c8c552f2f5b330d389336857da2c694234)