CMS-14336 Do not expose wicket internals
When issuing a GET request for one of the CMS' Wicket Ajax component URLs
Wicket returns a 500 error with a message:
  "component ${some/long/path} has been removed from the page".
This behaviour of echoing the component "path" has been flagged as a potential
security risk. We should change the CMS such that the response does not include
the component "path". The same applies to all WicketRuntimeExceptions that are
classified as internal server errors.
I decided to return the class name of the exception instance as the message so
that it is clear what the nature of the exception was without leaking sensitive
information.

(cherry picked from commit f74be005ea0a001405d864f1746031372690d0b6)