CMS-14460 [Backport 14.5] Log to be expected exceptions on INFO level
The ItemNotFoundException very easily happens in a clustered setup: Eg
on cluster node A do:

- save draft after changing some fields on descendant compounds
- commit editable instance
- obtain editable instance

The first step changes descendant nodes of the draft, the last step
replaces all the descendant nodes of the draft.

On cluster node B, the changes from the journal table of the save are
processed, but those nodes are also already being removed again by the
next changelog : during that process, the indexing of the nodes for the
first save typically hit an ItemNotFoundException: since the save,
commit, obtain scenario is very common, this happens frequently.

Therefore, log an info instead of a warn

Note that we should improve the typical commit -> obtain flow with a
workflow step which retains the draft, something like
commitAndRetainEditableInstance()

(cherry picked from commit 5c0b11662e303f0a3531583e8a94c8c630d9796f)