CMS-14401 [Backport 13.4] Add incremental suffix to duplicated doc type field paths
When a new field with an existing path is added to a document type,
an underscore(_) is added at the end of path to avoid of duplicate
paths. Now a numeric suffix will be added to the path names instead of
underscore(_).

For example;

myproject:boolean (existing field)

+ myproject:boolean (new field, will be updated as myproject:boolean1)

(cherry picked from commit 91d7a37af798764a88a9a49e734453401bd760a8)