CMS-14197 Prevent logout if there are unsaved document changes
If a user is editing a document in the right-side panel in the
Experience Manager then she/he is the holder of the document.
Nobody else can then edit the document. When a user tries to
switch to another perspective then the XM shows a modal dialog
where the user has to choose:
- cancel navigation and continue editing
- discard changes and navigate
- save changes and navigate
Choosing for discard or save will remove the user as the holder
of the document. When the user clicks the logout link we also
want make sure that document is released, so the same modal dialog
should be shown when the user clicks the logout link.

The navapp API already has a method beforeNavigation for this.
So use it to implement the desired behavior. Note that the CMS
is the parent of the XM, so we need to make sure that there also
is a beforeLogout method similar to beforeNavigation in the
navapp-bridge.js.

(cherry picked from commit d5a7e637b895e6e64829246942a4ba3919193f21)