CMS-14356 [Backport 14.6] Remove "dirty" check from CKEditor form-submit event
When saving a document, we use a form-submit event to update the
hidden <textarea> field of a CKEditor field with the latest value.
This <textarea> element is then used to sent data to the server.

For performance reasons, we only update the <textarea> if the editor
considers itself dirty, but this wont work for certain use-cases, as
the dirty check will compare the current value against the *initial*
value.

Consider the following use-case:

- start with an empty & required HTML field
- press save, now the editor is marked invalid (because it is required)
- add some text and press save, now is is marked valid
- remove the text again and press save

At this point, the field should be marked invalid again, but because
the current value is equal to the *initial* value (both empty) the
editor considers itself *not* dirty and we don't update the <textarea>.

Removing the dirty check might introduce a small regression in
performance, but considering that it *only* happens when we save a
document, the user should not be affected by it.