ENT-2509 Open the right page instead of home
Open the right page instead of home when using the view menu of the
content editor for any other branch than "Core".

Set the renderPathInfo in HippoIframeService#load because that path is
used to create the path in HippoIframeService#reload.

In CMS_14141 #0c62e8c5 HippoIframeService#getCurrentPathInfo() is used
in HippoIframeService#reload. Before HippoIframeService#getSrc() was
used.
HippoIframeService#reload creates a path based on
HippoIframeService#getCurrentPathInfo. The renderPathInfo at that point
was "undefined" and the ChannelService#makePath returns the content root.
So the home page was shown.

No that the renderPathInfo has been set, the right page is shown.

(cherry picked from commit 85a25ba938aae1a1d46c395f28440d7890423031)