CMS-14408 use a FencedFeedbackPanel
If a login fails then a feedback message is added to the
feedback component of the Wicket Session. If the user then
logs in successfully the Wicket Session already exists and
still contains the error message. When a new (default) Wicket
FeedbackPanel is created it will include feedback messages of
all of it's ancestors, so also the messages of the Session
feedback. A dialog however should in general only show feedback
for it's own child components. Wicket provides the class
FencedFeedbackPanel just for this case. So use it to prevent
parent feedback from showing up in the modal dialog.

Note: I could also have 'fixed' it by clearing all the feedback
messages In the PluginUserSession just before trying to login.
I think that's still a good idea, but I prefer this fix first
because it already does the job.

(cherry picked from commit 6b2be7f80e2253f519bbe614b2fb1f1970f72d80)