CMS-14556 [Backport 14.6] Support Campaign workflow on documents
On a handle, when the mixin hippo:versionInfo is present, the handle
can contain hippo:versionMeta property which can contain a json string
representing a VersionsMeta object. For now, the VersionsMeta object
only contains Campaign objects for specific versions, but after this,
we'll add support for having a 'label' for a version to be stored as
well. The reason for using a json object is that there will be more data
to store in the future, and storing this kind of data in multi-values
props on a handle is really unhandy, and at the same time, storing the
data on child nodes below a handle is too much of a performance overhead
and introduces unneeded complexity. The json usage is not super nice,
there are no JCR query capabilities now, but at least it is very
flexible and suited to be expanded