CMS-14557 [Backport 14.6] Support serving the campaign version of a document
In case a document is requested by the object converter, then regardless
whether this document is an XPage Document or a normal document,
regardless whether the document is the primary document for the current
URL or just some linked document, make sure that in case there can be
found an active campaign matching the request branch (project branch),
then serve that active campaign. If multiple active campaigns match,
take the one with the shortest duration.

Note that the implementation introduced a DocumentCampaignService : this
has the advantage that the implementation can be in hst platform, which
can have a dependency on the utility class JcrVersionsMetaUtils in the
repository workflow. Also note that this exact same service is needed
in the Channel Manager content-service later since we need it to be able
to flag the (historic) version which is active : a flag we need to add
to the versions returned from the /versions endpoint

(cherry picked from commit ca545f5f3d2b452fa460bdbccb8f40b76bdd417d)