CMS-14594 [Bacport 14.6] For 'versions' endpoint, add extra flags to indicate state
Globally for all versions there is now a new flag:

isLive = true | false

this indicates whether the document is available in live website

Next to that, a version has a flag:

published = true | false

It indicates which version corresponds to the published variant in case
the published variant is also 'live'. In general a published live
variant always corresponds to a version which has {branchId}-published
label. However only for bootstrapped live documents this is not always
true if they are never published through document workflow. Hence, in
that case, there is some extra logic which version corresponds to the
published live

Lastly, a flag has been added which is

active = true | false

This indicates which version would actually be served by the HST live
rendering. This can be a different one than the published variant
version in case of a document campaign which is running. Also note
that 'active' can be true while isLive = false : in this case it
indicates which version would be served in case the document gets
published (though during publication, a new version is created and as
a result, the 'active' moves to the latest added version in case there
is no campaign

(cherry picked from commit ea751d435985a78ba3aeb461896811cd12d14a3f)