CMS-14685 [Backport 14.7] Expose if "maxitems" has been configured on a field
By default, a field that is a multiple has a max-values of
Integer.MAX_VALUE. Although this is a restriction, it is very unlikely
that a field ever contains this many values.

When the max-values is set to the default value, the frontend should
not display a message like "X of maximum Y". To make that decision,
we expose a new property "hasMaxValues" which indicates if the
"maxitems" property has been configured on the field.

(cherry picked from commit bfadde4358dd1e1af21c2409e6be6f1ca9ac13a5)