CMS-14685 [Backport 14.7] Remove gab between choice&compound fields
The vertical blue border in the VisualEditor indicates which fields
belong to the same compound field.

Between the compound fields of the same parent, there is now a 12px
gab which should be removed to make it more clear that these fields
belong to the same parent.

(cherry picked from commit d753f4325d927176bbeec33d1ba8448565879c52)