CMS-14950 Keep the draft before switching the editor
Some content type definitions contain fields that cannot be edited in the
Experience Manager. If you try to create such a document then in step 2
you can't save it because not all required fields can be filled in.
In that case a link is displayed that opens the draft in the content editor
so that the unsupported fields can be filled in and the document can be saved.
The user might already have filled in some fields in step 2 in the Experience
Manager. With the introduction of instant preview these values live in the
in-memory session only, but since there is no upublished variant yet the in-memory
model cannot be saved. To prevent losing the values of these fields we now have to
call saveDraftIfDirty before switching to the Content Editor.

(cherry picked from commit 738ce0f1660d5d8fa212b81c91b8c4c257b58d7a)