ENT-2941 Reset client apps on site selection
It used to be the case we would reuse apps that did not support site
selection. This was wrong however as there potentially could be an
account that does not actually have access to any of the brxm navItems
and therefore should not load that client app. Also it could be possible
that we will have other client apps in the future that will not support
site selection (such as Exponea) but could not be accessible by a
certain acount.

(cherry picked from commit 5bfc6307fd05b38da6d80c3572c1b5af7702228a)