CMS-14806 Mark 'content-service' interactions to be CM rest requests
To be able to support this, we first had to support in/out interceptors
for subclasses of the JsonResourceServiceModule.

It was not yet possible in subclasses of JsonResourceServiceModule to
register in/out interceptors. That has been added to the
JsonResourceServiceModule

After that, make sure that every request to the 'content-service' gets
intercepted to mark it with

ChannelManagerRestRequest = true