CMS-14809 [Backport 14.6] When a campaign is running, xpage doc creation should work
At this moment, when trying to create an XPage document in the *CONTENT
EDITOR*, eg for the channel with id 'foo', then this would work, until
a branch (project) of 'foo' would be started as campaign. The problem
is very delicate:

1) The XPageLayouts are tried to be fetch for channel id 'foo' in
ChannelServiceImpl#getXPageLayouts(channelId). This contains a filter
which tries to find the channel with id 'foo'. This happens with

channelId.equals(mount.getChannel().getId())

2) *However*, what needs to be realized, is that mount.getChannel()
returns the channel for the 'current context' in case there are
branches. Thus that code for example in the CM returns the channel
for branch A if that branch is selected. The DelegatingHstSiteProvider
delegates depending on the context to the channelManagerHstSiteProvider
or to the websiteHstSiteProvider to return the right HstSite from the
CompositeHstSite (composite is the master + branches)

3) *However*, the DelegatingHstSiteProvider falls back to the
websiteHstSiteProvider in case the request is not a channel mgr preview
or channel mgr rest request. But the websiteHstSiteProvider does return
the branch in case a campaign is live: As a result, not 'foo' is
returned by mount.getChannel().getId() but someting like foo-v34G6 which
does not match the request channelId

4) As a result, in the end, ChannelServiceImpl#getXPageLayouts(channelId)
returns an empty collection, resulting in the new xpage document popup to
only show 'new document' and a 'x' to close it, nothing else

The fix is to introduce a separate state, namely 'contextLess' for the
use case where the request originates from eg wicket some management API