CMS-14519 Swallow unexpected exceptions
Swallow unexpected exceptions in production. An error will be logged
and in case of a wicket javascript error the error will be shown in
the javascript console of the browser.

Rename the NavAppExceptionMapper to SwallowExceptionMapper, because
the exceptions are not communicated to the navigation application
anymore.

In wicket development mode an error page with stacktrace is still
shown.

The NavAppExceptionMapping called the navigation application to show a
toast in case of an unexpected exception. From a ux perspective that
is undesirable, because the unexpected error does not contain an
action or information that is of immediate relevance for the user.

Swallow the unexcepted exceptions by using a EmptyRequestHandler in
combination with the SHOW_EXCEPTION_PAGE display type [1], because
wicket either shows an error page or re-throws the exception so the
servlet container will show an error page.

[1] https://ci.apache.org/projects/wicket/apidocs/8.x/org/apache/wicket/settings/ExceptionSettings.html#SHOW_NO_EXCEPTION_PAGE

(cherry picked from commit a6855c8a20a22563acf204a343cf2ab388bf7378)