CMS-14727 Await resolve of all images
When replacing the DOM nodes of a component entity (container &
container-item), we wait before all images contained in the component
entity have been loaded by the browser.

Our code worked fine for well behaving images, but both broken and
unknown images resulted in a frozen UI since we did not propertly detect
these.

Next to that, I think our code did not property wait for *all*
images to load, since it invoked the "onLoadCallback" for every
successfully loaded image.

To fix it, I changed our implementation by looking at how
https://github.com/fregante/image-promise works. First, it adds a second
"error" listener for properly detecting broken images and then it uses
the "image.naturalWidth" property to detect if the browser actually
considers the image to have a size. This seems to be the most widely
used approach for this use-case (also mentioned a lot on Stackkoverflow)

Note: I had to use the "window.$Promise" reference to be able to use $q
instead of native Promises to make the tests pass.