CMS-14727 Await page-reload when adding component with new-head-contributions
When a component is added to a container, we call
PageStructureService.renderContainer(). This function checks if the
component contains new header contributions (like style elements), and
if so, it reloads the page. This is done by emitting an event named
"page:new-head-contributions" (we can't access the HippoIframeService
directly as it introduces circular refs).

When the page is reloaded, the communication service is first destroyed
(on unload of the iframe) and then re-created (on load of the iframe).

Because renderContainer() does not wait for the page to be reloaded,
it returns to ComponentCatalogService._addComponent which will try
to load the new component and open the side-panel to edit it.
Since the communication service is destroyed at this point, we run into
errors.

The solution is to add a  callback fn to the emit of event
"page:new-head-contributions" and invoke it after the page has
finished reloading.

Note:  PageStructureService.renderComponent() also uses the same
event to handle the same case (new-head-contributions) but it does not
have to wait for the page to be reloaded, so I did not change that.