CMS-14842 Prevent non-whitelisted nodes
Throw an exception if a non-whitelisted element or attribute is found in
the xml.

Use the SAX parser, because the scaling of svg's also uses the SAX
parser. If we would be lenient towards malformed xml during the upload
of the gallery item, the scaling would still fail.

The owasp library is non present in any of our used maven repositories,
the SAX parser is part of the java core. So the SAX parser has preference.

(cherry picked from commit 0041769a406e89b5405871a235b24fa9bd00d76c)