CMS-14842 Detect javascript url in style element
The only advantage of using the owasp html sanitizer over plain SAX is
that the owasp library can also detect javascript url in the style tag.

It throws a SAX exception in that case. The vanilla user feedback is
left, because that's out of scope for this issue.

(cherry picked from commit f26b23744481213f7c8263ed47679a6dc93b39b3)