CMS-14855 Detect javascript url in style element
The only advantage of using the owasp html sanitizer over plain SAX is
that the owasp library can also detect javascript url in the style tag.

It throws a SAX exception in that case. The vanilla user feedback is
left, because that's out of scope for this issue.

(cherry picked from commit dccc2f0badd885af8deaadac13e382b1a6caa8ee)