CMS-14842 Detect javascript url in style element
The only advantage of using the owasp html sanitizer over plain SAX is
that the owasp library can also detect javascript url in the style tag.

It throws a SAX exception in that case. The vanilla user feedback is
left, because that's out of scope for this issue.

(cherry picked from commit e4e8de2a819d08987f249367556118fbab3c1e42)