CMS-14833 Use Tika for content detection for assets and images
Instead of relying on a very old single class ImageInfo which is only
capable of detecting the mimetype based on its actual contents of a very
limited set of image extensions, we now rely on Tika which is able to
detect the mimetype of an extremely large set of file extensions.

As a result, I've deprecated (for BC since api) the ImageInfo class and
the method validateMimeType in MimeTypeHelper since it should not be
used any more.

The following extensions are tested and supported, but most likely more
are supported: JPG,PNG,GIF,TIFF,ICO,SVG,WEBP,DOC, DOCX,XLS, XLSX,PPT,
PPTX,PDF,ODT,ODS,ODP,RTF,MP3,WAV,OGG,AVI,MOV,MP4,OGG,WMV,WEBM,CSV,JSON,
XML,HTML,ZIP

EXE has been tested and is not supported to be uploaded: not explicitly
but Tika identifies an EXE differently than the browser provided
mimetype: so this is not an explicit exclusion. Perhaps we should make
it explicit.

When renaming file its extension, then in apart from some exceptions,
the upload will never be allowed and a feedback is given wrt mimetype:
if the file extension is changed, it is most likely not a valid file
to be uploaded. Only exception is renaming .jpg to .jpeg.

Note that as a result of this commit, the property configuration for
mimetypes.allowed has become completely redundant. Customers who already
have configured the property won't suffer any downside : the real meaning
of the property was to skip certain extensions for validation because
the old ancient ImageInfo could not handle those. We now handle validation
for all extensions we can think off.

If mimetypes.allowed is configured, we'll log a warning that the property
can be removed as it is not used any more.