CMS-14835 Clean up UPDATED/RESOLVED status
Clean up lingering UPDATED/RESOLVED status. The Dutch translations for
the affected keys are in line with the English translations.

In CHANNELMGR-2230 we registered translations and already added Dutch
translations. We have not defined that in our process. As a result
during the integration of these translations for 13.2.0, the status
UPDATED/RESOLVED was (accidentally) introduced. The translation was
updated and the Dutch translation had been resolved ( prior to sending
it to the translation agency ).