CMS-14833 Support extension to detected allowed mimetype mapping
Depending on the browser, some file extensions are uploaded by the
browser with giving it a mimetype which is not equal to the one detected
by Tika and neither a sub or super type. As a result, the file is not
allowed for upload.

The feedback in the browser is something like:

The file 'sample4.aac' content does not match with the content MIME
type: audio/x-aac

(note only the English translation is changed, other languages say
'given MIME type' which is not a problem since clear enough)

If a customer really wants to support .aac uploads then, and for example
it doesn't work for chrome, then it is easy to support this via
configuration: At

/hippo:configuration/hippo:frontend/cms/cms-services/assetValidationService

you can now configure the multi-valued property
'extension.mimetype.allowed.mappings' which can support failing types.
In the case of the message above, to 'extension.mimetype.allowed.mappings'
they only need to add the allowed mapping:

.aac,audio/x-aac

This makes .aac also pass while still being checked for the content, so
you can't change, say, a .exe to .aac and then upload it.

Note that changing the 'extension.mimetype.allowed.mappings' in a running
environment requires a new logout and login of the cms user to 'see' the
config changes.

Lastly, wrt implementation details: Unfortunately, the CMS FE hardly
works with real global services exposed via the HippoServiceRegistry:
that would make it much easier since I'd want to be able to use the
'extension.mimetype.allowed.mappings' config in the MagicMimeTypeFileItem
ideally, however, that would require building now a full mime type
configuration service accessible in MagicMimeTypeFileItem. Since this is
really too much work, I resorted to using unfortunately a thread local
to make sure that ONLY during validation phase, the getContentType
extra checks in MagicMimeTypeFileItem kicks in: if we also do it after
the validation phase, the file upload later on fails since it would
throw an InvalidMimeTypeException on getContentType later