CMS-14852 Consistently wait after jcr session saves before continuing
[backport release/14.7]

A set of tests in the page-composer, most notably
CRUDPageAndModifyContainerTest fail frequently. This is due to JCR async
behavior wrt when other sessions see each others changes. The only
simple 'fix' for this is to make sure after a session.save some waiting
is done. In 75% of the places we already do this, but in quite some
places we do not do this properly

(cherry picked from commit d34db7fac955e6a1caa53000b70b208f409bc720)