CMS-15128 [Backport v14.7] Do not use shared system session but instead workflow session
When using a shared system session, there is concurrency on the session
which is not allowed.

Also it seemed to me that the changes where done by the session backed
by the 'handle' node which thus would result in no changes in the
system session at all. If instead the session backed by the handled
would be saved, it would result in issues for non-admins: A typical
editor / author is not allowed to write to the unpublished or published
variants but only to the draft! This is by ACL

Therefore instead of relyin on the user session or system session, use
the backing workflow internal workflowSession which is PER user and thus
save to use (no concurrency) and which also is allowed to write changes
to the  unpublished and published variants

Note the unit test still has to be adjusted

(cherry picked from commit 466f47d50d5dcd5fd5c47f7e1818b25bc3354703)
(cherry picked from commit 6003ca3685823ab0c2855b2dab85184745f7fb57)