CMS-14942 [Backport v15.0] Don't break editing a document when an error occurs
The DocumentValidityService should not break the current behavior,
which means that it should never prevent the user from editing a
document. Hence, we don't throw InternalServerErrorException but log
an error or warning instead and simply return or continue to process
the next field.

Currently, a draft variant is required, so we also check for that.

(cherry picked from commit e8b5488224b8e9996c45ab3de7fb73bb8c11d307)