CMS-14928 [Backport 15.0] Sync overlay after lazy-load images are loaded
First, we now wait for all non lazy-load images to finish loading,
before we redraw the overlay when the page changes. This prevents
broken overlays when rendering large images (non-cached) over slow
connections.

The code already existed in the component-entity and is moved into a
utils module, so it can be shared.

To work around lazy-load images, a second callback was introduced that
will be invoked once the browser decides it's time to show the image.
In this case the main promise will be resolved immediately as not to block
the overlay rendering.

(cherry picked from commit 29ca2f1fb893bc76ae04b6931e4053ca61003872)