CMS-14950 session delegation support is mandatory for instant preview
For instant preview to work, session delegation support is mandatory so
remove the option to disable it. It is super unlikely that any customer
ever disabled it. Disabling is not supported any more in 15.0. We do not
need to document since extremely unlikely that it is used.

If we do would not support session delegation, the changes potentially
present in the 'commands' would never be updated on the preview session
and as a result would not be visible to the preview session

By removing this support it will now always work for the instant preview

Next to that, checking the preview session for being pristine is awkward
since it will frequently not be pristine when there are 'instant preview
changes'. Hence pending changes are just expected. They are however not
expected on the cms user session. Therefore changed to only check the
cms user session for being pristine