CMS-14951 Add command for node addition
The branch information and the construction of the ContentTypeHandle are
not needed for on-premise, because the content type are not publishable.

Lombok is not a dependency in the on-premise branch, so the
AddNodeFieldCommand is a delomboked version of the version on
release/saas.

(Compare with release/saas for a clear view of the differences)

Loosely based on ENT-3403