CMS-14950 Auto save draft
Backport auto save draft. This indeed is also part of the instant
preview functionality.

Compared to release/saas, only the auto save draft functionality only
applies to documents not to components ( That is part of the
X-Components feature) and the content type can be inferred without
branchId, because the document types on on-premise are not publishable (
there are no variants).

See ENT-3371

Thanks to Soumya for spotting.