CMS-14959 Implement new AmnesicPageStore
The storage of pages was reworked, so we needed to update the empty impl
of the AmnesicPageStore. I've overridden the "canBeAsynchronous" to
return true to suppress warnings from Wicket, but this will require some
testing.

Another change that has to be verified is that the
AjaxFileUploadBehavior "touched" a page af a successful upload and then
did "commitRequest", but that last method has been removed, so we need
to verify if calling "touchPage" is good enough.

(cherry picked from commit bcbb11ff1cb0aba4cab0bce0f4c32e97a3df8b4c)