CMS-14959 Replace Wicket Duration/Time with java Duration/Instant
Wicket9 has deprecated it's "time" package, so I've updated all
usage of the Wicket time classes to use the JDK's native classes.

One area that seems a bit problematic, is the parsing of a String value
into a "Duration". Previously, Wicket provided a very simple format to
do this: "4 minutes", "6 hours", etc, but this is not possible with
the native classes. So for this purpose, I've added the DurationUtils
which will try both methods and log a warning if the "old" way is used.

(cherry picked from commit 1552f8ac6ef85c72ee2a5fc900b0e23fe11985ca)