CMS-15128 [Backport v14.7] Backport DocumentValidityService API changes
The handleDocumentTypeChanges() is now called with a list of variants,
which is created in the DocumentService.

If any exception is thrown we check if the workflowSession still
has pending changes, and if so, discard them, so that we don't end up
with "locked" documents.

(cherry picked from commit 447abcc48c8fe377816d325694b35c0d77750e60)