CMS-15128 [Backport v14.7] Backport DocumentValidityService API changes
The handleDocumentTypeChanges() is now called with a list of variants,
which is created in the DocumentService.

If any exception is thrown we check if the workflowSession still
has pending changes, and if so, discard them, so that we don't end up
with "locked" documents.

(cherry picked from commit bc2c0ef08db77ae162558b8b2aeb84da8e74e460)