CMS-14984 Skip any dxphst properties from the HstNode impls to be loaded
[Backport 15.0]

dxphst are never relevant to load in the HstModel and thus not in
HstNode instances. However, when they are present, they are loaded in
properties, and then for Menu Items these properties do get serialized
in the Page Delivery API response

Although a bit dirty to hardcode skipping properties for the dxphst
namespace in an upstream project (the HST), it is by far simplest and
effective.

(cherry picked from commit 5217b939f24f34f04e8110264f1ca8b683215a2b)