CMS-10563 Keep dialog open on enter in searchfield
Solve the following bug:
When a user presses <enter> in the search field in any of the picker
dialogs in the experience manager, the dialog closed.
The expected behavior should be that the list of documents, images or
assets is filtered and that the dialog does not close.

Root cause:
(1) The ButtonWrapper disables all keys ( except escape ) in any of the
input and html fields. ( See
org.hippoecm.frontend.dialog.ButtonWrapper#decorate and wicket.contrib.input.events.InputBehavior#getDisable_in_input ).
(2) The org.hippoecm.frontend.plugins.cms.widgets.SubmittingTextField
adds custom javascript, enter.js that prevents the <enter> event on the
form when <enter> is hit on the Textfield. It also adds an <enter>
listener on the TextField.
(3) When the form receives the <enter> event, the form is closed. ( See
org.hippoecm.frontend.dialog.AbstractDialog#handleSubmit ).
(4) When a user pressed <enter> and the form receives the event at
first, the event is handled by both the SubmittingTextField and the
form, so the form closes. ( and the list of images, documents or assets
is updated ).
(5) When the TextField receives the <enter> event first, then the event
is handled by the SubmittingTextField, the event is prevented to reach
the form, so the form does not close and the the list of images,
documents or assets is updated.

We have not been able to determine, why the event reaches the
SubmittingTextField first in the Content perspective but not in the
Experience Manager. IMO depending on an event to reach a html element at
frist is just a bug.

Solution:
When the SubmittingTextField has focus, the handleSubmit of the form is
prevented.
Given the existing code, enabling/disabling or adding/removing the
<enter> event on the Ok button required too much code changes.
Preventing handleSubmit results has the same functional effect. When
inspecting the network tab in a browser however, still two callbacks are
made to the backend, one for the SubmittingTextField and one for the
form.

(cherry picked from commit 298f4671dcb5a31a31c2a18c145a14222bd0fcbd)
(cherry picked from commit 42536c8bcbaaf2b9051746bb9d6e2928ed86f009)
(cherry picked from commit 6c3d4cc2c19b842e69f989439a231b09e7d2ef6b)
(cherry picked from commit 638a1885208970fc09a936eb4d15c487c9858974)