CMS-14969 [Backport v15] Let the validity check process nested NodeFields
When we process a NodeFieldType, we should also validate the nested
fields of the NodeField. This requires us to look up the "DocumentType"
of the nested field, which is handled by the "DocumentTypesService" and
in turn requires a "UserContext". If a proper "DocumentType" is found,
we do a recursive call to "processChanges()".

(cherry picked from commit 30b5c606b69718a0c7acd46655d86c430a4e6dd3)