CMS-15009 Never serialize pages
As part of the changes to wicket 9, the Page store was reworked. The
initial implementation causes error to be logged that were caused by
serialization.

We don't serialize pages in our cms, because we don't use the pagestore
to keep track of versions of pages. Versions of pages are used by wicket
to enable back/forward functionality in the browser.

In our cms we only use two page versions, page 0 for the login screen
and page 1 for the screen with the navapp and the perspectives.

So we need a page store that only holds 2 pages and drops pages from the
store if serialization might be needed. That way the pages are never
serialized and the errors disappear.

(cherry picked from commit 5aa592eebc709e19835aa20c3a64f2e2303bee00)